Further progress on service classes refactoring.

+++ src/main/java/com/awoodcock/armory/service/ar/HandguardService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,8 +29,9 @@ public class HandguardService {
     }
 
     // READ
-    public Optional<Handguard> getHandguardById (int id) {
-        return handguardRepository.findById(id);
+    public Handguard getHandguardById (int id) {
+        Optional<Handguard> handguard = handguardRepository.findById(id);
+        return handguard.orElse(null);
     }
 
     public List<Handguard> getAllHandguards () {
@@ -37,9 +39,53 @@ public class HandguardService {
     }
 
     // UPDATE
+    public Handguard updateHandguardById (int id, Handguard handguard) {
+        Optional<Handguard> handguardOptional = handguardRepository.findById(id);
+
+        if (handguardOptional.isPresent()) {
+            Handguard originalHandguard = handguardOptional.get();
+
+            // name
+            if (Objects.nonNull(handguard.getName()) && !"".equalsIgnoreCase(handguard.getName())) {
+                originalHandguard.setName(handguard.getName());
+            }
+
+            // description
+            if (Objects.nonNull(handguard.getDescription()) && !"".equalsIgnoreCase(handguard.getName())) {
+                originalHandguard.setDescription(handguard.getDescription());
+            }
+
+            // price
+            if (Objects.nonNull(handguard.getPrice()) && handguard.getPrice() != 0) {
+                originalHandguard.setPrice(handguard.getPrice());
+            }
+
+            // weight
+            if (Objects.nonNull(handguard.getWeight()) && handguard.getWeight() != 0) {
+                originalHandguard.setWeight(handguard.getWeight());
+            }
+
+            // mountingSystem
+            if (Objects.nonNull(handguard.getMountingSystem()) && !"".equalsIgnoreCase(handguard.getMountingSystem())) {
+                originalHandguard.setMountingSystem(handguard.getMountingSystem());
+            }
+
+            // isFreeFloat
+            if (Objects.nonNull(handguard.isFreeFloat())) {
+                originalHandguard.setFreeFloat(handguard.isFreeFloat());
+            }
+
+            return handguardRepository.save(originalHandguard);
+        }
+        return null;
+    }
 
     // DELETE
-    public void deleteHandguard(int id) {
-        handguardRepository.deleteById(id);
+    public String deleteHandguard(int id) {
+        if (handguardRepository.findById(id).isPresent()) {
+            handguardRepository.deleteById(id);
+            return "Handguard successfully deleted.";
+        }
+        return "Handguard not found in database.";
     }
 }
